move request handler to standalone include
