+++ function/index.js
@@ -2,8 +2,7 @@
 
 const GoogleOAuthAuthorizer = require('./authorization');
 const GoogleDatastoreClient = require('./datastore');
-const { prefixedRejection } = require('./responseError');
-const validation = require('./validation.js');
+const toshandler = require('./toshandler.js');
 
 // handle CORS requests via 'cors' library
 const corsOptions = {
@@ -46,36 +45,8 @@ function tosapi(req, res, authClient, datastoreClient) {
     const authorizer = authClient || new GoogleOAuthAuthorizer();
     const datastore = datastoreClient || new GoogleDatastoreClient();
 
-    validation.validateRequestUrl(req);
-    validation.validateRequestMethod(req);
-    validation.validateContentType(req);
-    const authHeader = validation.requireAuthorizationHeader(req);
-    const reqinfo = validation.validateInputs(req);
-
-    return authorizer.authorize(authHeader)
-        .then(userinfo => {
-            if (req.method === 'GET') {
-                try {
-                    return datastore.getUserResponse(userinfo, reqinfo)
-                        .catch(err => {
-                            return prefixedRejection(err, 'Error reading user response');
-                        });
-                } catch (err) {
-                    return prefixedRejection(err, 'Error reading user response');
-                }
-            } else if (req.method === 'POST') {
-                try {
-                    return datastore.createUserResponse(userinfo, reqinfo)
-                        .catch(err => {
-                            return prefixedRejection(err, 'Error writing user response');
-                        });
-                } catch (err) {
-                    return prefixedRejection(err, 'Error writing user response');
-                }
-            } else {
-                return Promise.reject({statusCode: 405});
-            }
-        });
+    // TODO: route to different handlers for userRequest vs. status
+    return toshandler.handleRequest(req, authorizer, datastore);
 }
 
 /**
